conditional book button

+++ src/Pages/Dashboard/AllBuyers/AllBuyers.js
@@ -1,6 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import React from 'react';
 import toast from 'react-hot-toast';
+import Swal from 'sweetalert2';
 import Loader from '../../../Components/Loader/Loader';
 
 const AllBuyers = () => {
@@ -23,24 +24,44 @@ const AllBuyers = () => {
     }
 
     const handleDelete = id => {
-        fetch(`http://localhost:5000/buyerDelete/${id}`, {
-            method: 'DELETE',
-            headers: {
-                authorization: `bearer ${localStorage.getItem('secretToken')}`
+
+        Swal.fire({
+            title: 'Are you want to delete this?',
+            text: "It can't be undone!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Confirm'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/buyerDelete/${id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        authorization: `bearer ${localStorage.getItem('secretToken')}`
+                    }
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        if (data.success) {
+                            toast.success(data.message);
+                            refetch();
+                        }
+                        else {
+                            toast.error(data.message)
+                        }
+                    })
+
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
             }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.success) {
-                    toast.success(data.message);
-                    refetch();
-                }
-                else {
-                    toast.error(data.message)
-                }
-            })
     }
+
 
     return (
         <div className='my-8'>
